split ui functions across 2 files to lay the ground for fuzzing
